Now supports text layers
Pic2Print takes user imput in Preview/PostView email dialog, and passes
it to photoshop to be added as a text layer.  All processed images now
have a top layer added regardless if text is added.
Pic2Print also has a hex RGB input field on the config panel so the
operator can set the color of the text layer.