Version 10.01
Big change now is an addition to the printer definitions file to add a
scaling and offset to the actual prints.  This is needed because most
printers crop 1/8" off all sides.  Now, a scaling percentage and offset
allow the prints to be scaled and placed within the paper without any
cropping of the actual image.  These changes happend in -
1) printer.csv printer definition file. Each paper type has its own
scaling and offset parameters.
2) Pic2Print is modified to read this data and pass it to the config.txt
file when the config panel is closed.
3) automaticmode.js carries the code to receive the new parameters and
execute commands to photoshop for the scaling and offset.
4) Onsite.Printing.atn had the 2x6 photostrip action reduced since this
code now resides in javascript.
Minor other tweaks -
1)Post View window has been enlarged.
2)"Files to be loaded" scrolls correctly now.
3)Multiple background/foreground highlight is larger.