Version 11.05
Fixed, updated -
- fixed print paper size scaling. My code was scaling the image
up 1/8th of an inch. I thought the printers were doing this.
Updated the printer.csv file and automaticmode.js
- Formalized the time profiling in Pic2print and automatic.js so
the technician can profile the layout with filters.  This number
is entered in the Config panel in the printer profile times.
- In layout 000, a foreground layer of rulers is added for
measuring the effects of the scaling operations.
- Added three user defined filters, so the technician just has
to record these actions to add filters.
- In printers.csv, revised the scaling factors now that the .JS
code properly scales. No need to do it in this file.
- In Pic2Print, changed the Printer Time to Profile Time.
Also changed when the timers start on print countdowns to
make them more accurate. NOTE: Printer countdowns are
best estimates only, and used in the load balancing.