Version 11.07
New comic book filter uses halftoning.  Resetting the swatches now, so
the halftone does not accidentally invert.  Updated printers.csv for the
dnp ds40 for photo strip alignment. In the Javascript, I added what I
call a hack to fix a problem with 4x6,6x9,8x12 images. Seems resizing
the image leaves a few lines on the top & bottom, but only in this
ratio, don't know why it doesn't affect the other ratios yet.  The hack
will be removed soon..