Version 14.00
Major Changes:
1) Sped up the User UI for photobooth operation.  It was a bit slow in
transitions between frames.
2) The "Auto >>" checkbox on main control panel now displays and moves
to the right most image;
no need to click the Refresh button anymore if this box is checked. It
also updates the preview
window to reflect the new image.
3) A new checkbox was added to the main control panel called "Auto
View".  If checked it
will automatically display the preview or postview window depending on
the current activity.

Here's where it gets fun.

Check the three checkboxes, "Auto >>" and "Auto View" on the main
control panel, and the
"AutoScroll" on the postview window.  Open the preview window and
postview window on a
second display overlapping each other.  Use this display for the guests
to see their images
(you can experiment with this on one display). As images land in the
capture folder, they
will automatically be added to the list and the main window will scroll
to the far right to
display the new image. The preview panel will automatically open and
display the new image.
Now, click a number 1 - 10 to print the new image; the postview window
automatically pop up
to show the printed image. This goes back & forth to show the guests the
latest new image
and latest printed image.

Minor Changes:
1)The Preview button & Postview button text now change from lowercase
when hidden, to uppercase
when opened so the operator can see which window is currently displaying
on a second monitor
if the monitor is facing away.
2)In automaticmode.js,
A)Added code to change the document mode to RGB mode in case its
indexed, etc.
B)Removed an alert msg left in from prior work.
3)mmsForm.vb is completely depricated and can be deleted from the
project.
4)Fixed a bug in preview.vb if the email button is clicked.  It
referenced null data in
mmsForm.vb modules.
5)The email buttons on the Preview and Postview windows are now disabled
if email is disabled.
6)Removed the file name from the title bar in the Preview window.
7)Added a layout #007 for adding branding to images.