Version 14.08
HTML email has replaced the simple text email, so now you can build
custom branded emails for clients or sponsors.  See the email-read-me
file for instructions on how to format the html file.  The five
"email_imgXX.jpg" files are replaceable with your images, and can be
embedded in the emails.  The emailbody.html is an example email
demonstrating how to use all available images in your email.