Version 14.09
Fixed an outlook issue. Outlook keeps all inline images within a private
cache and doesn't allow these images to be saved.  So, the actual user
image has to be sent as an attachment w/o a CID or else outlook swallows
the image.  Now the user image can be saved as intended.  The
emailbody.html file had to drop using the user image in the email.  The
image can only ride as an attachment, not a usable image in the email
body.