Version 14.10
Email changes to fix txtmsgs receiving too large of images.  So, if an
email is sent to the phone, the alternate smaller image is sent rather
than the actual printed image.  If and email address and phone # are
give, the email will receive this small image too.  Along the way, I
found the html email attachements were broken, so this was fixed too.